Show the user when a deck is already added
This took a while to figure out because Observers don't force property closures
to generate more observables. This makes sense in retrospect. The only
documentation I could find on this was in this ticket:

    https://github.com/mobxjs/mobx/issues/1142

You have to import a special `Observer` component and wrap it around any
components generated in the property closures.