Final touches to card layout
Use simple flexbox attributes (space-around) instead of inserting
extra view with a flex value to add space.

Dynamically determine the max number of lines in the headword by
counting newlines in the text.

Allow the example sentence component to expand as needed, but cut it
off at the bottom of the card so that it doesn't overflow.

Handle display of example sentence component when there is missing data.

Factor out common code into methods.