Better styling for swipe labels
Generate colors for good/bad/okay swiping. I'm the least sure about the choice
of mahogany for the "bad" swipe, since it's a little bit orange. Oh well.

Cover the entire background of a card with the good/bad/ok color. Use icons
instead of unicode chars, which for some reason ignored the text color
attribute. Move the card layout style into its own file because it needs to be
shared between the Card component and the swipe label container component,
which is an Animated.View created by the swiper library and therefore cannot
be replaced.