Improved overlay opacity animation
The defaults for opacity animation in react-native-deck-swiper cause the overlay
to appear quite suddenly, so the effect of the opacity transition is almost
lost. Change the settings so that the opacity transition happens over a longer
distance. Also, don't make the card see-through. It's better for indicating the
meaning of a swipe because it makes the good/OK/bad overlay much clearer.