Basic audio working
This sucked to implement because assets have to be loaded at compile time,
so we have to hard-code the paths to audio that are to be used in `require`
statements. I'm also not entirely sure I want to stick with the expo library;
even though it is super simple to use, it may be better down the line to use
react-native-video, especially considering possible reuse for an ASL deck.