Working basic jest test for AddDeckButton
Whew! It was a lot of work getting here. Expo icon fonts have to be pre-loaded,
expo-jest is older and doesn't allow the "//" key for comments, and there also
seems to be a problem where we can't use testRegex at all, since it assumes that
testMatch is also defined. This last one I didn't search for a ticket for, but
the other things have tickets.