Translate between generating function of StudyManager and array of Swiper
It is pretty difficult for me to think of the app in terms of a simple growing
array of cards, since we dynamically add more and may schedule different groups
of cards in an interlaced pattern and the current card index is stuck inside the
swiper component.

It may be a little dirty or somehow overkill, but no we just translate our model
(a generating function called `getNextCard`) into the one used by Swiper (the
cards array) using a proxy. The implementation is longer than I though because
we have to include lookahead logic to account for the always-shown stack of
three cards.