TypeScript with Jest working
This was really difficult to figure out. For whatever reason, we have to use
just the transform from ts-jest's `jsWithBabel` preset; we can't use the
`defaults` preset like they show in the docs, and we can't use everything from
the preset like they show in the docs. Just the transform. ¯\_(ツ)_/¯