Workaround for bad types until jest-expo is released next. See https://github.com/expo/expo/pull/4520.
I thought I could just not touch the node_modules/jest-expo directory, but
yarn keeps overwriting it when I do other stuff.

I will have to keep this repo in mind:
https://github.com/dai-shi/typescript-expo-apollo-boilerplate. Seems to be
keeping up to date with the quirks of using expo/jest/typescript, and I got the
idea for this fix from there (though it's a different fix).