Add delete button for first time cards
This is just the UI logic. For now it just prints a homestarrunner reference :)

https://www.youtube.com/watch?v=7rrZ-sA4FQc

Also, note that we don't need to know the exact size of a button and set the
border radius to half of that to get a circle. We just set the border radius
to basically infinite and no matter the size of the button it will always be a
circle.