Working UI for deletion with confirmation
I found it a little bit weird to write because the react-native-modal component
requires rendering it once and showing/hiding it by setting a property. Since I
also wanted to change all of the text and button handlers in it, I would rather
have rendered and unrendered it. But I tried that, and I couldn't
programmatically animate the thing away. At least this way it looks better than
with the default modal :)