Redo the Card type
Change to an interface and use all methods instead of fields. Rename to ICard
to make TS happy. Allow for multiple user languages when retrieving the user
language headword and example sentence.

The RxDocument method `getAttachment`'s type did not indicate that it could
return null, but actually it can! So that case needs to be handled. Sent a PR
for that: https://github.com/pubkey/rxdb/pull/1306