Update StudyManager to use Scheduler card streams
Update StudyManager to use the streams returned by the CardSchedulers to 1)
keep track of the number of cards that are due for study and to 2) create an
observable array of cards to display via the Swiper.

It was a bit complex to figure this out because we are now working with RxJS,
Mobx, Promises and even Optionals! The way this is done, it should all be
concurrency safe. `mergeMap` takes care of waiting for promises to resolve and
then calling the subscribers' next() method sequentially.

I'm not entirely sure if we need to keep so many class variables around. Plenty
of `TODO` notes added!