`Promise.reject()` leads to untrackable errors!
I temporarily added the `bluebird` library, which has better trackability for
Promises, and found out that `Promise.reject()` was giving us the mystery error.
`reject(...)` should take an `Error` object as the argument so that we get a
message, a stack trace, etc.

I would like to leave the bluebird library in there, but it's giving more
mystery errors like the following:

    %cWarning: a promise was created in a handler but was not returned from it, see http://goo.gl/rRqMUw
        (No stack trace)
    From previous event:
        getItem@http://127.0.0.1:19001/node_modules/expo/AppEntry.bundle?platform=ios&dev=true&minify=false&hot=false:91208:25
        get@http://127.0.0.1:19001/node_modules/expo/AppEntry.bundle?platform=ios&dev=true&minify=false&hot=false:244034:38
    From previous event:
        get@http://127.0.0.1:19001/node_modules/expo/AppEntry.bundle?platform=ios&dev=true&minify=false&hot=false:244034:48
        _default@http://127.0.0.1:19001/node_modules/expo/AppEntry.bundle?platform=ios&dev=true&minify=false&hot=false:244357:19
        _callTimer@http://127.0.0.1:19001/node_modules/expo/AppEntry.bundle?platform=ios&dev=true&minify=false&hot=false:27865:17
        _callImmediatesPass@http://127.0.0.1:19001/node_modules/expo/AppEntry.bundle?platform=ios&dev=true&minify=false&hot=false:27901:19

To me this looks like something to do with the app bundling, which I absolutely
CANNOT change at all. Maybe Expo will fix this sometime? There's no stack trace
anyway, so it's completely unactionable and it pollutes the console logs. If I
need bluebird again, I can uncomment the shim code in polyfill.js.