Copy sqlite DB into couchDB and tweak the data structure a little
Add the `png` extension to the image attachments because not having an extension
is really confusing.

Add a `pronunciation` field to the cards, since it's not always possible to
automatically generate a pronunciation key (without creating another database
XD).

Create a new script for copying the ThaiPod101 data from the sqlite db into
couchDB.